requirements.txt
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.3
django-allauth==0.38.0
django-braces==1.13.0
django-heroku==0.3.1
django-oauth-plus==2.2.9
django-oauth-toolkit==1.2.0
django-oauth2-provider==0.2.6.1
django-rest-auth==0.9.3
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djangorestframework-oauth==1.1.0
gunicorn==19.9.0
httplib2==0.12.0
idna==2.8
Jinja2==2.10
livereload==2.6.0
Markdown==3.0.1
MarkupSafe==1.1.0
mkdocs==1.0.4
oauth2==1.9.0.post1
oauthlib==2.1.0
Pillow==5.3.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycparser==2.19
PyJWT==1.7.0
python-decouple==3.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.7
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.0.0
shortuuid==0.5.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
tornado==5.1.1
urllib3==1.24.1
whitenoise==4.1.2
